[Fix] refactor source code tree
