+++ arch/x86/kernel/init.cpp
@@ -1,0 +1,10 @@
+#include <arch.hpp>
+#include <kapi/gdt.hpp>
+#include <kapi/idt.hpp>
+
+bool arch_init() {
+    gdt_init();
+    idt_init();
+
+    return true;
+}
