+++ include/arch.hpp
@@ -1,0 +1,6 @@
+#ifndef ARCH_HPP_
+#define ARCH_HPP_
+
+bool arch_init();   // Returns false on error
+
+#endif  // ARCH_HPP_
